changed components name to account
