--- src/pages/dashboard/index.js
@@ -1,1 +1,0 @@
-export * from './Dashboard'
